Created Service implementation with interface and redefined the file stucture
